pyproject.toml
[project]
name = "stt_server_test"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "vhzkclq0705", email = "vhzkclq0705@naver.com"},
]
dependencies = ["fastapi>=0.115.12", "uvicorn[standard]>=0.34.2", "python-multipart>=0.0.20", "faster-whisper>=1.1.1", "psutil>=7.0.0", "boto3>=1.38.10", "kafka-python>=2.2.4"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
